refactoring, update TODO
