+++ app/botmanager.ts
@@ -4,10 +4,10 @@ import { Piscina } from 'piscina';
 import { fileURLToPath } from 'url';
 import { MessageChannel } from 'worker_threads';
 
-import manager_controller_ee from './eventemitters/manager-controller.eventemitter.ts';
-
 import bot_config_json from './configs/bot.config.json';
 import webdriver_config_json from './configs/webdriver.config.json';
+
+import manager_controller_ee from './eventemitters/manager-controller.eventemitter.ts';
 
 import { AIConfig, WorkerConfig } from './interfaces/config.interface.ts';
