+++ app/routes/bot.route.ts
@@ -1,5 +1,5 @@
 import * as express from 'express';
-import * as bot_controller from '../controllers/bot-controller.ts';
+import * as bot_controller from '../controllers/bot.controller.ts';
 
 const router = express.Router();
